requirements.txt
flask
gunicorn
Pillow
pytest